+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * repr of `Map` and `Set`
 * add `digit-value` function
 * add `#d` prefix for numbers
+* add `include` macro
 ### Bugfix
 * fix load file that have shebang
 * fix repr of eof object